% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_legis_returns.R
\name{clean_legis_returns}
\alias{clean_legis_returns}
\title{A description.}
\usage{
clean_legis_returns(dir, is_precinct = TRUE)
}
\arguments{
\item{dir}{A directory}
}
\value{
A dataframe
}
\description{
A more detailed escription.
}
