% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_nm_election_data.R
\name{clean_nm_election_data}
\alias{clean_nm_election_data}
\title{A description.}
\usage{
clean_nm_election_data(dir, is_precinct = TRUE)
}
\arguments{
\item{dir}{A directory}
}
\value{
A dataframe
}
\description{
A more detailed escription.
}
